Download Processed Citizen Voting Age Population Data

Downloads processed CVAP data for a state. CVAP data is rounded to the nearest 5 so totals may not sum properly, but will be close.

cvap_get(
  state,
  year = 2019,
  geography = "block group",
  out_file = NULL,
  moe = FALSE,
  clean = TRUE
)

Arguments

state
character. The state to get data for or nation for the nation file.

year
numeric. Year for the data in 2009 to 2019.

geography
character. Level of geography. Default is 'block group'. See Details.

out_file
file to save downloaded rds to

moe
Include margin of error? Default is FALSE.

clean
Should variable names be standardized? Default is TRUE.

Value

tibble of data

Details

Geography options are

'block group': block group level data

'cd': congressional district data (years 2016+)

'county': county-level data

'place': place-level data

shd': state house district data (years 2016+)

'ssd': state senate district data (years 2016+)

'state': state-level data

'tract': tract-level data

'nation': nation-wide data

Examples

cvap_get('DE')
#> # A tibble: 574 × 10
#>    GEOID     cvap cvap_white cvap_black cvap_hisp cvap_asian cvap_aian cvap_nhpi
#>    <chr>    <dbl>      <dbl>      <dbl>     <dbl>      <dbl>     <dbl>     <dbl>
#>  1 1000104…  2425       2135        130        55         60         0         0
#>  2 1000104…  1575       1430         90        20         15         0         4
#>  3 1000104…  1225        945        210         0         25        50         0
#>  4 1000104…  1940       1460        265       125         35        20         0
#>  5 1000104…  1445        905        490        30          0         4         0
#>  6 1000104…  4470       2710       1475        85         40         4         0
#>  7 1000104…  7770       5090       2150       285         80         4         0
#>  8 1000104…  2815       1235       1425       125          0         0         0
#>  9 1000104…  1140       1045         40        55          0         0         0
#> 10 1000104…  1675       1255        300       105          4         0         0
#> # … with 564 more rows, and 2 more variables: cvap_two <dbl>, cvap_other <dbl>